+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/cloudformation/stack/OutputActionPopup.kt
@@ -6,6 +6,7 @@ package software.aws.toolkits.jetbrains.services.cloudformation.stack
 import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.actionSystem.DefaultActionGroup
 import com.intellij.ui.PopupHandler
+import software.aws.toolkits.jetbrains.ui.CopyAction
 import software.aws.toolkits.resources.message
 import java.awt.Component
 
@@ -26,4 +27,3 @@ internal class OutputActionPopup(private val selected: () -> SelectedOutput?) : 
 }
 
 internal data class SelectedOutput(internal val key: String?, internal val value: String?, internal val outputName: String?)
-
