+++ jetbrains-core/src/software/aws/toolkits/jetbrains/ui/CopyAction.kt
@@ -1,0 +1,21 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package software.aws.toolkits.jetbrains.ui
+
+import com.intellij.icons.AllIcons
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.ide.CopyPasteManager
+import com.intellij.openapi.project.DumbAware
+import java.awt.datatransfer.StringSelection
+
+class CopyAction(name: String, private val value: String?) : AnAction(name, null, AllIcons.Actions.Copy), DumbAware {
+    override fun update(e: AnActionEvent) {
+        e.presentation.isEnabledAndVisible = value != null
+    }
+
+    override fun actionPerformed(e: AnActionEvent) {
+        CopyPasteManager.getInstance().setContents(StringSelection(value))
+    }
+}
